Guess commentable status
For a thin command line client with basic POST capability, it would useful to
let the software take control over whether the post or comment is commentable
or not.

This does not stop people froom creating commentable comments, but the system
cannot handle it in threads.

Also, make sure a category is given, because that is the whole point of
sorting entries.